initial version

+++ src/main/java/io/rentalapp/common/DateRangeDetails.java
@@ -9,13 +9,16 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Class to hold the date related details of the rental agreement
+ */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class DateRangeDetails {
-    Long totalWeekDays = Long.valueOf(0);
-    Long totalWeekendDays  = Long.valueOf(0);
-    Long totalHolidays = Long.valueOf(0);
+    Integer totalWeekDays = Integer.valueOf(0);
+    Integer totalWeekendDays  = Integer.valueOf(0);
+    Integer totalHolidays = Integer.valueOf(0);
     List<LocalDate> dateRange = new ArrayList<LocalDate>();
 }
